step2/backend/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
PyMySQL==1.1.1
python-dotenv==1.1.0
urllib3==1.25.10
WooCommerce==3.0.0
gunicorn